wip: rt drawing -- converting uvb self-shielding density to code unit

+++ src/rhyme_uv_background/src/rhyme_uv_background_get.f90
@@ -8,14 +8,14 @@ module function rhyme_uv_background_get(uvb, z, species, logger) result(rates)
    character(len=*), dimension(:), intent(in) :: species
    type(logger_t), intent(inout) :: logger
 
-   real(kind=4), dimension(2*size(species)) :: rates
+   real(kind=8), dimension(2*size(species)) :: rates
 
    select case (uvb%model)
    case (uvbid%HM12)
       rates = rhyme_uv_background_haardt_madau_12_get(z, species)
    case default
       call logger%err('Unknown UVB model!', 'model', '=', [uvb%model])
-      rates = 0e0
+      rates = 0d0
    end select
 end function rhyme_uv_background_get
 end submodule get_smod
